+++ aragora/nomic/dev_coordination/core.py
@@ -1,23 +1,20 @@
-"""Development coordination primitives for concurrent multi-agent work.
+"""Core implementation of ``aragora.nomic.dev_coordination``.
 
-This module adds the missing control plane for high-churn concurrent work:
-- Work leases with explicit write scopes and expected tests
-- Completion receipts for bounded worker outputs
-- Integration decisions for an explicit integrator lane
-- Salvage candidates for dirty worktrees and stashes
+Contains the SQLite-backed :class:`DevCoordinationStore`, the work-order
+classification jungle, scope/glob helpers, and the argparse CLI.
 
-The design intentionally builds on existing Aragora orchestration patterns:
-- EventBus for cross-worktree signaling
-- GlobalWorkQueue-compatible work item projection
-- Receipt-style content hashes for auditability
-- Git-common-dir local state so agents coordinate without tracked-file churn
+Split out of the original 5311-line ``dev_coordination.py`` module as
+TCP-3 PR-A.  Dataclasses/enums/errors now live in
+:mod:`aragora.nomic.dev_coordination.models`, and stateless helpers live
+in :mod:`aragora.nomic.dev_coordination.utils`.  The public import
+surface is preserved verbatim through ``__init__.py`` re-exports.
 """
 
 from __future__ import annotations
 
 import argparse
 import asyncio
-import hashlib
+import hashlib  # noqa: F401 — re-exported as ``aragora.nomic.dev_coordination.hashlib`` for ``dev_salvage``
 import json
 import os
 import re
@@ -25,11 +22,9 @@ import shlex
 import sqlite3
 import subprocess
 import uuid
-from dataclasses import dataclass, field
-from datetime import datetime, timedelta, timezone
-from enum import Enum
+from datetime import datetime, timedelta
 from pathlib import Path, PurePosixPath
-from typing import TYPE_CHECKING, Any
+from typing import Any
 
 from aragora.docs_only import (
     canonical_docs_container_scope,
@@ -38,15 +33,30 @@ from aragora.docs_only import (
     is_docs_safe_top_level_file,
 )
 from aragora.nomic import dev_coordination_verification as _verification_helpers
+from aragora.nomic.dev_coordination.models import (
+    CompletionReceipt,
+    DeveloperTask,
+    IntegrationDecision,
+    IntegrationDecisionType,
+    LeaseConflictError,
+    LeaseStatus,
+    SalvageCandidate,
+    SalvageStatus,
+    WorkLease,
+)
+from aragora.nomic.dev_coordination.utils import (
+    _has_wildcard,
+    _json_dump,
+    _json_loads,
+    _normalize_claim,
+    _parse_dt,
+    _safe_kill_probe,
+    _utcnow,
+)
 from aragora.nomic.event_bus import EventBus
-from aragora.nomic.global_work_queue import GlobalWorkQueue, WorkItem, WorkStatus, WorkType
+from aragora.nomic.global_work_queue import GlobalWorkQueue, WorkItem, WorkStatus
 from aragora.worktree.fleet import FleetCoordinationStore
 
-UTC = timezone.utc
-if TYPE_CHECKING:
-    from aragora.swarm.lane_telemetry import LaneTelemetryCollector
-
-_LANE_TELEMETRY: LaneTelemetryCollector | None = None
 _ACTIVE_LEASE_STATUSES = {"active"}
 _PENDING_INTEGRATION_DECISIONS = {"pending_review"}
 _OPEN_SALVAGE_STATUSES = {"detected", "claimed"}
@@ -111,15 +121,6 @@ def _is_docs_only_path(path: Any) -> bool:
     return normalized.startswith("docs/") or normalized.endswith((".md", ".mdx", ".rst", ".txt"))
 
 
-def _get_lane_telemetry() -> LaneTelemetryCollector:
-    global _LANE_TELEMETRY
-    if _LANE_TELEMETRY is None:
-        from aragora.swarm.lane_telemetry import LaneTelemetryCollector
-
-        _LANE_TELEMETRY = LaneTelemetryCollector()
-    return _LANE_TELEMETRY
-
-
 def _normalize_completion_outcome(
     *,
     outcome: str,
@@ -134,440 +135,6 @@ def _normalize_completion_outcome(
         return normalized
     has_deliverable = bool(pr_url.strip() or pr_number is not None or commit_shas or changed_paths)
     return "deliverable_created" if has_deliverable else "clean_exit_no_deliverable"
-
-
-class LeaseConflictError(ValueError):
-    """Raised when a lease overlaps another active lease."""
-
-    def __init__(self, conflicts: list[dict[str, Any]]):
-        super().__init__("Lease overlaps existing active work")
-        self.conflicts = conflicts
-
-
-class FileScopeViolationError(ValueError):
-    """Raised when a completion touches files outside the claimed scope."""
-
-    def __init__(self, violations: list[dict[str, Any]]):
-        super().__init__("Completion violates file-scope ownership")
-        self.violations = violations
-
-
-class LeaseStatus(str, Enum):
-    """Lifecycle states for work leases."""
-
-    ACTIVE = "active"
-    COMPLETED = "completed"
-    RELEASED = "released"
-    EXPIRED = "expired"
-
-
-class IntegrationDecisionType(str, Enum):
-    """Integrator verdict for completed work."""
-
-    PENDING_REVIEW = "pending_review"
-    MERGE = "merge"
-    CHERRY_PICK = "cherry_pick"
-    REQUEST_CHANGES = "request_changes"
-    DISCARD = "discard"
-    SALVAGE = "salvage"
-
-
-class SalvageStatus(str, Enum):
-    """Lifecycle states for salvage candidates."""
-
-    DETECTED = "detected"
-    CLAIMED = "claimed"
-    PORTED = "ported"
-    DISCARDED = "discarded"
-
-
-@dataclass(slots=True)
-class WorkLease:
-    """A bounded claim over a task, worktree, and write scope."""
-
-    lease_id: str
-    task_id: str
-    title: str
-    owner_agent: str
-    owner_session_id: str
-    branch: str
-    worktree_path: str
-    allowed_globs: list[str] = field(default_factory=list)
-    claimed_paths: list[str] = field(default_factory=list)
-    expected_tests: list[str] = field(default_factory=list)
-    status: str = LeaseStatus.ACTIVE.value
-    created_at: str = field(default_factory=lambda: _utcnow().isoformat())
-    updated_at: str = field(default_factory=lambda: _utcnow().isoformat())
-    expires_at: str = field(default_factory=lambda: (_utcnow() + timedelta(hours=8)).isoformat())
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-    @property
-    def is_active(self) -> bool:
-        return self.status in _ACTIVE_LEASE_STATUSES and not self.is_expired
-
-    @property
-    def is_expired(self) -> bool:
-        return _parse_dt(self.expires_at) <= _utcnow()
-
-    def overlaps(self, allowed_globs: list[str], claimed_paths: list[str]) -> bool:
-        other_globs = [_normalize_claim(x) for x in allowed_globs if str(x).strip()]
-        other_paths = [_normalize_claim(x) for x in claimed_paths if str(x).strip()]
-        if self.claimed_paths and _claims_overlap(self.claimed_paths, other_globs, other_paths):
-            return True
-        return _globs_overlap_any(self.allowed_globs, other_globs, self.claimed_paths, other_paths)
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "lease_id": self.lease_id,
-            "task_id": self.task_id,
-            "title": self.title,
-            "owner_agent": self.owner_agent,
-            "owner_session_id": self.owner_session_id,
-            "branch": self.branch,
-            "worktree_path": self.worktree_path,
-            "allowed_globs": list(self.allowed_globs),
-            "claimed_paths": list(self.claimed_paths),
-            "expected_tests": list(self.expected_tests),
-            "status": self.status,
-            "created_at": self.created_at,
-            "updated_at": self.updated_at,
-            "expires_at": self.expires_at,
-            "metadata": dict(self.metadata),
-        }
-
-    @classmethod
-    def from_row(cls, row: sqlite3.Row) -> WorkLease:
-        return cls(
-            lease_id=row["lease_id"],
-            task_id=row["task_id"],
-            title=row["title"],
-            owner_agent=row["owner_agent"],
-            owner_session_id=row["owner_session_id"],
-            branch=row["branch"],
-            worktree_path=row["worktree_path"],
-            allowed_globs=_json_loads(row["allowed_globs_json"], []),
-            claimed_paths=_json_loads(row["claimed_paths_json"], []),
-            expected_tests=_json_loads(row["expected_tests_json"], []),
-            status=row["status"],
-            created_at=row["created_at"],
-            updated_at=row["updated_at"],
-            expires_at=row["expires_at"],
-            metadata=_json_loads(row["metadata_json"], {}),
-        )
-
-
-@dataclass(slots=True)
-class CompletionReceipt:
-    """A bounded worker output ready for integration review."""
-
-    receipt_id: str
-    lease_id: str
-    task_id: str
-    owner_agent: str
-    owner_session_id: str
-    branch: str
-    worktree_path: str
-    base_sha: str = ""
-    head_sha: str = ""
-    commit_shas: list[str] = field(default_factory=list)
-    changed_paths: list[str] = field(default_factory=list)
-    tests_run: list[str] = field(default_factory=list)
-    validations_run: list[str] = field(default_factory=list)
-    assumptions: list[str] = field(default_factory=list)
-    blockers: list[str] = field(default_factory=list)
-    outcome: str = "completed"
-    risks: list[str] = field(default_factory=list)
-    pr_url: str = ""
-    pr_number: int | None = None
-    confidence: float = 0.0
-    created_at: str = field(default_factory=lambda: _utcnow().isoformat())
-    artifact_hash: str = ""
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-    def __post_init__(self) -> None:
-        if not self.validations_run and self.tests_run:
-            self.validations_run = list(self.tests_run)
-        if not self.artifact_hash:
-            self.artifact_hash = _artifact_hash(
-                {
-                    "lease_id": self.lease_id,
-                    "task_id": self.task_id,
-                    "owner_agent": self.owner_agent,
-                    "owner_session_id": self.owner_session_id,
-                    "branch": self.branch,
-                    "worktree_path": self.worktree_path,
-                    "base_sha": self.base_sha,
-                    "head_sha": self.head_sha,
-                    "commit_shas": self.commit_shas,
-                    "changed_paths": self.changed_paths,
-                    "tests_run": self.tests_run,
-                    "validations_run": self.validations_run,
-                    "assumptions": self.assumptions,
-                    "blockers": self.blockers,
-                    "outcome": self.outcome,
-                    "risks": self.risks,
-                    "pr_url": self.pr_url,
-                    "pr_number": self.pr_number,
-                    "confidence": self.confidence,
-                    "metadata": self.metadata,
-                }
-            )
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "receipt_id": self.receipt_id,
-            "lease_id": self.lease_id,
-            "task_id": self.task_id,
-            "owner_agent": self.owner_agent,
-            "owner_session_id": self.owner_session_id,
-            "branch": self.branch,
-            "worktree_path": self.worktree_path,
-            "base_sha": self.base_sha,
-            "head_sha": self.head_sha,
-            "commit_shas": list(self.commit_shas),
-            "changed_paths": list(self.changed_paths),
-            "tests_run": list(self.tests_run),
-            "validations_run": list(self.validations_run),
-            "assumptions": list(self.assumptions),
-            "blockers": list(self.blockers),
-            "outcome": self.outcome,
-            "risks": list(self.risks),
-            "pr_url": self.pr_url or None,
-            "pr_number": self.pr_number,
-            "confidence": self.confidence,
-            "created_at": self.created_at,
-            "artifact_hash": self.artifact_hash,
-            "metadata": dict(self.metadata),
-        }
-
-    @classmethod
-    def from_row(cls, row: sqlite3.Row) -> CompletionReceipt:
-        return cls(
-            receipt_id=row["receipt_id"],
-            lease_id=row["lease_id"],
-            task_id=row["task_id"],
-            owner_agent=row["owner_agent"],
-            owner_session_id=row["owner_session_id"],
-            branch=row["branch"],
-            worktree_path=row["worktree_path"],
-            base_sha=row["base_sha"],
-            head_sha=row["head_sha"],
-            commit_shas=_json_loads(row["commit_shas_json"], []),
-            changed_paths=_json_loads(row["changed_paths_json"], []),
-            tests_run=_json_loads(row["tests_run_json"], []),
-            validations_run=_json_loads(row["validations_run_json"], []),
-            assumptions=_json_loads(row["assumptions_json"], []),
-            blockers=_json_loads(row["blockers_json"], []),
-            outcome=row["outcome"],
-            risks=_json_loads(row["risks_json"], []),
-            pr_url=row["pr_url"],
-            pr_number=row["pr_number"],
-            confidence=float(row["confidence"]),
-            created_at=row["created_at"],
-            artifact_hash=row["artifact_hash"],
-            metadata=_json_loads(row["metadata_json"], {}),
-        )
-
-
-@dataclass(slots=True)
-class DeveloperTask:
-    """Canonical task-queue projection for supervised swarm work."""
-
-    task_key: str
-    task_id: str
-    run_id: str
-    goal: str
-    title: str
-    status: str
-    priority: int = 50
-    owner_agent: str = ""
-    reviewer_agent: str = ""
-    blocked_by: list[str] = field(default_factory=list)
-    acceptance_checks: list[str] = field(default_factory=list)
-    allowed_paths: list[str] = field(default_factory=list)
-    lease_id: str | None = None
-    owner_session_id: str | None = None
-    branch: str | None = None
-    worktree_path: str | None = None
-    receipt_id: str | None = None
-    created_at: str = field(default_factory=lambda: _utcnow().isoformat())
-    updated_at: str = field(default_factory=lambda: _utcnow().isoformat())
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "task_key": self.task_key,
-            "task_id": self.task_id,
-            "run_id": self.run_id,
-            "goal": self.goal,
-            "title": self.title,
-            "status": self.status,
-            "priority": self.priority,
-            "owner_agent": self.owner_agent or None,
-            "reviewer_agent": self.reviewer_agent or None,
-            "blocked_by": list(self.blocked_by),
-            "acceptance_checks": list(self.acceptance_checks),
-            "allowed_paths": list(self.allowed_paths),
-            "lease_id": self.lease_id,
-            "owner_session_id": self.owner_session_id,
-            "branch": self.branch,
-            "worktree_path": self.worktree_path,
-            "receipt_id": self.receipt_id,
-            "created_at": self.created_at,
-            "updated_at": self.updated_at,
-            "metadata": dict(self.metadata),
-        }
-
-    def to_work_item(self) -> WorkItem:
-        status = _developer_task_work_status(self.status)
-        blockers = list(self.blocked_by)
-        if status == WorkStatus.BLOCKED and not blockers:
-            blockers = [f"task_status:{self.status}"]
-        return WorkItem(
-            id=f"task:{self.task_key}",
-            work_type=WorkType.CUSTOM,
-            title=self.title,
-            description=self.goal or self.title,
-            status=status,
-            created_at=_parse_dt(self.created_at),
-            updated_at=_parse_dt(self.updated_at),
-            base_priority=self.priority,
-            assigned_to=self.owner_session_id or self.owner_agent or None,
-            blockers=blockers,
-            tags=["developer-task", "swarm", self.status],
-            metadata=self.to_dict(),
-        )
-
-
-@dataclass(slots=True)
-class IntegrationDecision:
-    """Integrator verdict for a completion receipt."""
-
-    decision_id: str
-    lease_id: str
-    receipt_id: str
-    decision: str
-    target_branch: str
-    rationale: str
-    chosen_commits: list[str] = field(default_factory=list)
-    followups: list[str] = field(default_factory=list)
-    decided_by: str = ""
-    created_at: str = field(default_factory=lambda: _utcnow().isoformat())
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "decision_id": self.decision_id,
-            "lease_id": self.lease_id,
-            "receipt_id": self.receipt_id,
-            "decision": self.decision,
-            "target_branch": self.target_branch,
-            "rationale": self.rationale,
-            "chosen_commits": list(self.chosen_commits),
-            "followups": list(self.followups),
-            "decided_by": self.decided_by,
-            "created_at": self.created_at,
-        }
-
-    def to_work_item(self) -> WorkItem:
-        priority = 85 if self.decision == IntegrationDecisionType.PENDING_REVIEW.value else 50
-        return WorkItem(
-            id=f"integration:{self.decision_id}",
-            work_type=WorkType.CUSTOM,
-            title=f"Integration review for receipt {self.receipt_id[:8]}",
-            description=self.rationale or f"{self.decision} for lease {self.lease_id}",
-            status=WorkStatus.READY,
-            created_at=_parse_dt(self.created_at),
-            updated_at=_parse_dt(self.created_at),
-            base_priority=priority,
-            tags=["integration", self.decision, self.target_branch],
-            metadata=self.to_dict(),
-        )
-
-    @classmethod
-    def from_row(cls, row: sqlite3.Row) -> IntegrationDecision:
-        return cls(
-            decision_id=row["decision_id"],
-            lease_id=row["lease_id"],
-            receipt_id=row["receipt_id"],
-            decision=row["decision"],
-            target_branch=row["target_branch"],
-            rationale=row["rationale"],
-            chosen_commits=_json_loads(row["chosen_commits_json"], []),
-            followups=_json_loads(row["followups_json"], []),
-            decided_by=row["decided_by"],
-            created_at=row["created_at"],
-        )
-
-
-@dataclass(slots=True)
-class SalvageCandidate:
-    """Potentially useful abandoned work discovered in stashes or worktrees."""
-
-    candidate_id: str
-    source_kind: str
-    source_ref: str
-    branch: str = ""
-    worktree_path: str = ""
-    stash_ref: str = ""
-    head_sha: str = ""
-    changed_paths: list[str] = field(default_factory=list)
-    summary: str = ""
-    likely_value: float = 0.0
-    status: str = SalvageStatus.DETECTED.value
-    created_at: str = field(default_factory=lambda: _utcnow().isoformat())
-    updated_at: str = field(default_factory=lambda: _utcnow().isoformat())
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "candidate_id": self.candidate_id,
-            "source_kind": self.source_kind,
-            "source_ref": self.source_ref,
-            "branch": self.branch,
-            "worktree_path": self.worktree_path,
-            "stash_ref": self.stash_ref,
-            "head_sha": self.head_sha,
-            "changed_paths": list(self.changed_paths),
-            "summary": self.summary,
-            "likely_value": self.likely_value,
-            "status": self.status,
-            "created_at": self.created_at,
-            "updated_at": self.updated_at,
-            "metadata": dict(self.metadata),
-        }
-
-    def to_work_item(self) -> WorkItem:
-        return WorkItem(
-            id=f"salvage:{self.candidate_id}",
-            work_type=WorkType.MAINTENANCE,
-            title=f"Salvage {self.source_kind} {self.source_ref}",
-            description=self.summary or f"Review salvage candidate from {self.source_kind}",
-            status=WorkStatus.READY,
-            created_at=_parse_dt(self.created_at),
-            updated_at=_parse_dt(self.updated_at),
-            base_priority=max(10, min(100, int(self.likely_value * 100))),
-            tags=["salvage", self.source_kind, self.status],
-            metadata=self.to_dict(),
-        )
-
-    @classmethod
-    def from_row(cls, row: sqlite3.Row) -> SalvageCandidate:
-        return cls(
-            candidate_id=row["candidate_id"],
-            source_kind=row["source_kind"],
-            source_ref=row["source_ref"],
-            branch=row["branch"],
-            worktree_path=row["worktree_path"],
-            stash_ref=row["stash_ref"],
-            head_sha=row["head_sha"],
-            changed_paths=_json_loads(row["changed_paths_json"], []),
-            summary=row["summary"],
-            likely_value=float(row["likely_value"]),
-            status=row["status"],
-            created_at=row["created_at"],
-            updated_at=row["updated_at"],
-            metadata=_json_loads(row["metadata_json"], {}),
-        )
 
 
 class DevCoordinationStore:
@@ -4873,62 +4440,6 @@ def _developer_task_work_status(status: str) -> WorkStatus:
     return WorkStatus.PENDING
 
 
-def _safe_kill_probe(raw_pid: Any) -> Exception | None:
-    """Probe whether *raw_pid* is alive.  Returns ``None`` if alive, the exception otherwise."""
-    import os as _os
-
-    try:
-        _os.kill(int(raw_pid), 0)
-        return None
-    except (ProcessLookupError, PermissionError, TypeError, ValueError, OSError) as exc:
-        return exc
-
-
-def _utcnow() -> datetime:
-    return datetime.now(UTC)
-
-
-def _parse_dt(value: str) -> datetime:
-    return datetime.fromisoformat(value)
-
-
-def _json_dump(value: Any) -> str:
-    return json.dumps(_json_compatible(value), sort_keys=True)
-
-
-def _json_loads(value: str | None, default: Any) -> Any:
-    if not value:
-        return default
-    try:
-        return json.loads(value)
-    except json.JSONDecodeError:
-        return default
-
-
-def _artifact_hash(payload: dict[str, Any]) -> str:
-    return hashlib.sha256(
-        json.dumps(_json_compatible(payload), sort_keys=True).encode()
-    ).hexdigest()
-
-
-def _json_compatible(value: Any) -> Any:
-    if isinstance(value, bytes):
-        return value.decode("utf-8", errors="replace")
-    if isinstance(value, dict):
-        return {str(key): _json_compatible(nested) for key, nested in value.items()}
-    if isinstance(value, (list, tuple, set)):
-        return [_json_compatible(item) for item in value]
-    return value
-
-
-def _normalize_claim(value: str) -> str:
-    return value.strip().strip("/")
-
-
-def _has_wildcard(pattern: str) -> bool:
-    return any(token in pattern for token in ("*", "?", "["))
-
-
 def _path_matches_glob(path: str, pattern: str) -> bool:
     clean_path = _normalize_claim(path)
     clean_pattern = _normalize_claim(pattern)
@@ -4994,48 +4505,6 @@ def _claims_overlap(
             if _glob_overlap(claimed, other):
                 return True
     return False
-
-
-def _parse_worktree_entries(raw: str) -> list[tuple[Path, str]]:
-    entries: list[tuple[Path, str]] = []
-    current_path: Path | None = None
-    current_branch: str | None = None
-    for line in raw.splitlines():
-        text = line.strip()
-        if text.startswith("worktree "):
-            current_path = Path(text[len("worktree ") :]).resolve()
-            current_branch = None
-        elif text.startswith("branch refs/heads/"):
-            current_branch = text[len("branch refs/heads/") :]
-        elif text == "" and current_path is not None and current_branch is not None:
-            entries.append((current_path, current_branch))
-            current_path = None
-            current_branch = None
-    if current_path is not None and current_branch is not None:
-        entries.append((current_path, current_branch))
-    return entries
-
-
-def _status_paths(lines: list[str]) -> list[str]:
-    paths: list[str] = []
-    for line in lines:
-        text = line.strip()
-        if not text:
-            continue
-        path = text[3:] if len(text) > 3 else text
-        if " -> " in path:
-            path = path.split(" -> ", 1)[1]
-        paths.append(_normalize_claim(path))
-    return paths
-
-
-def _estimate_salvage_value(*, ahead: int, changed_paths: list[str], dirty: bool) -> float:
-    value = 0.2
-    if dirty:
-        value += 0.2
-    value += min(0.4, ahead * 0.1)
-    value += min(0.2, len(set(changed_paths)) * 0.02)
-    return max(0.0, min(1.0, value))
 
 
 def _build_parser() -> argparse.ArgumentParser:
